Added a Refresh entry in the PullDownMenu.
This Refresh button triggers the data update.
Also added a "isRefreshing" property in Controller that is exposed to QML. This one allows us to make the PullDownMenu blink, disable the Refresh menu entry when updating, etc.
Also exposed the date and time of the last update (needs a proper fix).
Also added some other stuff that was needed for the previous ones to work.