Updated CoverPage.
The cover displays the status of the parking lots marked as Favorites.
You can cycle through them via a CoverAction.

If there are no Favorites, the cover displays a placeholder that invites the user to add some (not implemented yet).
You can also refresh the data via another CoverAction (not implemented yet).